backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.21
asyncpg==0.28.0
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
structlog==23.1.0
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1 